Move all stories into the stories directory
