+++ src/implementations/react/src/stories/Checkbox.story.js
@@ -19,7 +19,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text, boolean } from '@storybook/addon-knobs';
 
-import { default as Checkbox } from '';
+import { default as Checkbox } from '../adapters/FormElements/CheckboxAdapter';
 
 storiesOf('Checkbox', module)
   .addWithInfo('Basic checkbox', '', () => {
