+++ src/implementations/react/src/stories/Dropdown.story.js
@@ -3,7 +3,7 @@ import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import { text, select, boolean } from "@storybook/addon-knobs";
 
-import Dropdown from "./DropdownAdapter";
+import Dropdown from "../adapters/FormElements/DropdownAdapter";
 
 const Option = Dropdown.Option;
 
@@ -98,4 +98,4 @@ storiesOf("Dropdown", module)
       </Dropdown>
     );
   });
-	
+
