+++ src/implementations/react/src/stories/GlobalNav.story.js
@@ -19,7 +19,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { select, boolean, text } from '@storybook/addon-knobs';
 
-import GlobalNav from './GlobalNavAdapter';
+import GlobalNav from '../adapters/GlobalNav/GlobalNavAdapter';
 
 const SideNav = GlobalNav.SideNav;
 const TopNav = GlobalNav.TopNav;
@@ -109,12 +109,12 @@ const LONG_COPY = (
   </div>
 );
 
-import logo from '../../playground/images/bim-logo.png';
-import profileImage from '../../playground/images/profileImage.png';
-import project1 from '../../playground/images/project-1.png';
-import project2 from '../../playground/images/project-2.png';
-import project3 from '../../playground/images/project-3.png';
-import project4 from '../../playground/images/project-4.png';
+import logo from '../playground/images/bim-logo.png';
+import profileImage from '../playground/images/profileImage.png';
+import project1 from '../playground/images/project-1.png';
+import project2 from '../playground/images/project-2.png';
+import project3 from '../playground/images/project-3.png';
+import project4 from '../playground/images/project-4.png';
 
 storiesOf('GlobalNav', module)
   .addWithInfo('default', <div><p>Global Nav basic usage</p></div>, () => {
