+++ src/implementations/react/src/stories/IconButton.story.js
@@ -19,7 +19,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text } from '@storybook/addon-knobs';
 
-import IconButton from './IconButton';
+import IconButton from '../adapters/IconButtonAdapter';
 
 storiesOf('IconButton', module)
   .addWithInfo('Icon Button', ``, () => {
