+++ src/implementations/react/src/stories/Profile.story.js
@@ -18,10 +18,10 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text, boolean } from '@storybook/addon-knobs';
-import GlobalNav from '../../../../adapters/GlobalNav/GlobalNavAdapter';
+import GlobalNav from '../adapters/GlobalNav/GlobalNavAdapter';
 
-import logo from '../../../../playground/images/bim-logo.png';
-import profileImage from '../../../../playground/images/profileImage.png';
+import logo from '../playground/images/bim-logo.png';
+import profileImage from '../playground/images/profileImage.png';
 
 const TopNav = GlobalNav.TopNav;
 const Profile = TopNav.Profile;
