+++ src/implementations/react/src/stories/Search.story.js
@@ -18,8 +18,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text } from '@storybook/addon-knobs';
-import GlobalNav from '../../../../adapters/GlobalNav/GlobalNavAdapter';
-import logo from '../../../../playground/images/bim-logo.png';
+import GlobalNav from '../adapters/GlobalNav/GlobalNavAdapter';
+import logo from '../playground/images/bim-logo.png';
 
 const TopNav = GlobalNav.TopNav;
 const TopNavSearch = TopNav.Search;
