+++ src/implementations/react/src/stories/TextField.story.js
@@ -19,7 +19,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { text, select, boolean } from '@storybook/addon-knobs';
 
-import { TextField } from '../../../react-hig';
+import { TextField } from '../react-hig';
 
 storiesOf('TextField', module)
   .addWithInfo('By default', ``, () => {
