+++ src/implementations/react/src/stories/index.js
@@ -1,22 +1,22 @@
 
 import 'hig-vanilla/dist/hig.css';
 
-import '../elements/basics/Button.story';
-import '../elements/basics/IconButton.story';
-import '../elements/components/FormElements/Checkbox.story';
-import '../adapters/FormElements/PasswordFieldAdapter.story';
-import '../adapters/FormElements/RangeAdapter.story';
-import '../adapters/FormElements/RadioButtonAdapter.story';
-import '../adapters/SpacerAdapter.story';
-import '../adapters/FormElements/TextAreaAdapter.story';
-import '../elements/basics/form-elements/TextField.story';
-import '../elements/components/GlobalNav/TopNav/Search.story.js';
-import '../elements/components/GlobalNav/TopNav/Profile.story';
-import '../adapters/GlobalNav/GlobalNavAdapter.story';
-import '../adapters/GlobalNav/TopNav/ProjectAccountSwitcher.story';
-import '../elements/components/GlobalNav/SubNav/Tabs.story';
-import '../elements/components/GlobalNav/SideNav/SideNav.story';
-import '../elements/components/Modal.story';
-import "../adapters/FormElements/DropdownAdapter.story";
-import '../elements/components/Typography.story';
-import '../adapters/TableAdapter.story.js';
+import './Button.story';
+import './Checkbox.story';
+import "./Dropdown.story";
+import './GlobalNav.story';
+import './IconButton.story';
+import './Modal.story';
+import './PasswordField.story';
+import './Profile.story';
+import './ProjectAccountSwitcher.story';
+import './RadioButton.story';
+import './Range.story';
+import './Search.story.js';
+import './SideNav.story';
+import './Spacer.story';
+import './Tabs.story';
+import './Table.story.js';
+import './TextArea.story';
+import './TextField.story';
+import './Typography.story';
